Add mechanism for "generic" commands. This permits to specify which arguments of a command have to be translated.
Here are some examples:
A module can use:
register_generic("addcontentsline,0,3,,3");
to specify that the \addcontentsline command has 3 arguments (no optional)
and only the third one has to be translated (i.e. neither the file nor the
section).

An user can use in its input file:
% po4a: command cite 1,1,1,
to specify that the cite command has one optional argument followed by a
mandatory argument, and only the optional argument (if present) has to be
translated (i.e. not the citation key).