There were a problem with the added/removed non-breaking spaces. This could be a general issue (not only for non-breaking spaces).
This was triggered by a French translation containing a "à" (agrave).
When the output document is UTF-8, this character is encoded as 0xC3 0xA0,
and this last 0xA0 was changed to "\ " in post_trans.

This fix consists in using the non-breaking spaces from the input document
charset (with a fall back to a non probable string) before the translation.
After the translation, we change the non-breaking spaces to the groff one
only if the output charset allows non-breaking spaces.