Fix 02 test
The problem is that the same msgid appeared twice with the
same reference; this reference was written twice in the PO
file, but is written only once now.
This may be a change in gettext or in po4a, but well, who
cares?
Change one msgid to avoid duplicate references.