Move po4a-build call from Makefile to Build.PL
This is a work in progress; in po4a 0.41, I want to
remove the Makefile and do not let Build.PL call
po/bin/Makefile, see
  https://bugzilla.redhat.com/show_bug.cgi?id=629931

The proper way to let xgettext deal with different
languages is to call xgettext once for each language
and concatenate POT files.

Symlinking po4a-build.sh is no more needed.