porefs: Add a wrap/nowrap optional specifier
By default, po4a does not wrap file locations.  This is a bad idea,
because gettext tools do wrap them.  But if we modify this behavior,
most generated PO files are modified, which will cause trouble
when developers with different versions of po4a make changes.
So the --porefs has been extended, its syntax is now
  (full|none|noline)(,(wrap|nowrap))?
Default is "full,nowrap".  In some months (years?), we will change
this default to "full,wrap".