AsciiDoc: fix two line titles when encoding is different from ascii
Two line titles consist of a title line followed by an underline, which
has the same width as the title line.  We thus have to compute this
width, which requires knowing text encoding.
But text comes either from PO files or master documents.  A simple
solution is to always take encoding from PO file, and modify po4a-gettextize
to artificially set $transtt->{TT}{po_in}{encoder}.

Text width is computed by the Unicode::GCString module.  If this module
is not available and encoding is not ASCII, processing stops.