AsciiDoc: add support for automatic comments
AsciiDoc has two kinds of comments:
 1. line comments, which consists of "//"
 2. line block, enclosed by lines of slashes (at least 4 slashes)

Those comments will be extracted and written into PO files; they may
be useful to give hints to translators.

There is one caveat though, it is not always clear to know if comments
must be written before or after the text.
Normally comments are written just before text.  But this is different
with lists.  For instance consider a list like:

  // Comment 1
  - Note 1.
  // Comment 2
  - Note 2.
  // Comment 3

Running AsciiDoc parser with 'showcomments' attribute shows that comment 2
belongs to the first item, and comment 3 to the second item.
So here po4a will put comments 1 and 2 into the first item, and comment 3
into the second one.

Since comment location is quite hard to parse correctly, they are not
written into translated files in order to avoid confusion.