AsciiDoc: Fix the handling of literal paragraphs, again (fix #9)
It's tough to detect that a paragraph in a list block should be marked
as verbatim. That's a paragraph that is more indented than it should.
But the paragraphs that are attached to a labeled list (enumerate or
such) *must* be indented to be detected.

So, in my previous fix, I excluded the labeled lists (for which
self->{bullet} is defined) and checked if the block is indented
(length($indent)>0).

This is not enough because in the git documentation, they put verbatim
blocks in labeled lists. It cannot be the first block under the label
(can it?), and it is detected because the second (or more) block is
further indented than the previous one. Thus the new test to detect if
we should unwrap the block: the indentation increased since last block.

There is some false positive to this test, as blocks in nested list:
item::
  blablabla
  subitem;;
        This paragraph is detected as no-wrap where it should not.
	I hope that this is harmless.

Lastly, if a block in a list is no-wrap, then any subsequent blocks of
that list were wrongly marked as no-wrap too. This is not the case
anymore.