tests: kill t/compare-po.pl
This script was used to compare PO files, but the diagnosis were a
nightmare to debug: It only gave the first changed line without any
context. I changed the tests to use diff -u instead:

  diff -u "-I'Copyright (C) 20.. Free Software Foundation, Inc.' -I'^# Automatically generated, 20...' -I'^\"POT-Creation-Date:' -I'^\"PO-Revision-Date:'

That's not great looking, but you just have to write PODIFF in the
test command lines to get it.