asciidoc: compat option to match strictly equal block fences
If a block contains a line of the same kind of characters but of
different length, this line is interpreted as a closing fence, just
like python asciidoc.

This is an issue because this means that no such line can be included
in a block. Asciidoctor solves that issue by requiring the opening and
closing fences to be the same length, which can be chosen arbitrarily
greater than 3.

The 'compat=asciidoctor' option makes the same test.

Fixes #239