Regenerate the potfile if it's empty, even if it's recent
The documented "quick start" is to simply touch the file to give its
name, so don't ask the user to touch a file of the doc afterward.

Also, add a test to ensure that this scenario is working as expected.

Also, update the doc to not hint about the possibility that the doc
should be touched before proper use. It's just adding useless noise
here.