Request the a semantic of perl that is only 10 years old
v5.12 was released in 2012 while v5.6 was released in 2000.

The reason for the change is to be safe wrt the Unicode Bug, where the
semantic of operations over utf strings is harder to predict. The file
encodings are really a pain in po4a and I'd like to clean this part if
possible.