Fix the --translate-only option
I added a lot of verbose messages to understand what's going on,
rewrote a bit of code I didn't understant but was not recently
modified, and the bug introduced last month was gone.

At the end, I'm not sure of what happened, but this
fixes https://github.com/mquinson/po4a/issues/482

Thanks Alexander for the diagnostic and the reproducer!