Try to not fail when writing UTF po files with old Perl versions
This is some sort of blind attempt, but I'm rather confident
Fix https://github.com/mquinson/po4a/issues/495

I still cannot reproduce this bug locally, but Ineiev tracked it down
and tested the patch proposed with Perl v5.16.3 v5.24.1 v5.26.1
v5.26.3 v5.28.1 v5.32.1 v5.34.0 v5.36.0 v5.36.3 v5.38.2. Many thanks
to them about it!

That's a bit sad that it's not part of the v0.72 but duh, that's still
a blind attempt at fixing a problem I cannot reproduce. I'm getting
too old for the last-minute commit breaking the release. I think it
can wait, or maintainers can cherry-pick it if it works for them.