Set forgiving encoding fallback when parsing Po file
Before reading the expected encoding from the file itself, we first
try to read it as UTF-8 so we can at least get to read the line that
specifies the encoding.

For some reason, $PerlIO::encoding::fallback is set to FB_CROAK when
we reach this part of the code so the program would instead exit
at the first unknown character.

Unless we were requested a specific encoding, do not even print any
warning about it.