asciidoc: refactor the style handling
Style customization is possible in Asciidoc, and the formatting of
such new style can be defined in the configuration of the asciidoc
engine. This has an effect on how the styled paragraph or block is
supposed to be processed during the translation, mainly whether the
segment of text is wrapped or not (whether line breaks should be
preserved).

The style handling was a bit messy and making the verse style a
special case was not a good idea. This commit refactors the style
handling to make it more consistent and easier to understand.

This commit also introduces a new style definition extension with the
'%' character that allows to define a style of paragraphs and blocks
that is not wrapped.

fixes #548

Signed-off-by: Jean-Noël Avila <jn.avila@free.fr>